+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Unreleased]
+
+### Changed
+
+- new nodes are now initialised at centre of viewport instead of at origin
+
 ## [0.3.0] 2025-12-13
 
 ### Added